+++ web/WEB-INF/web.xml
@@ -2,15 +2,15 @@
 <web-app version="3.1" xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd">
     <servlet>
         <servlet-name>create</servlet-name>
-        <servlet-class>crud.tinySave</servlet-class>
+        <servlet-class>io.rerum.crud.tinySave</servlet-class>
     </servlet>
     <servlet>
         <servlet-name>update</servlet-name>
-        <servlet-class>crud.tinyUpdate</servlet-class>
+        <servlet-class>io.rerum.crud.tinyUpdate</servlet-class>
     </servlet>
     <servlet>
         <servlet-name>delete</servlet-name>
-        <servlet-class>crud.tinyDelete</servlet-class>
+        <servlet-class>io.rerum.crud.tinyDelete</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>create</servlet-name>
